Move variants chart to different folder
