+++ packages/app/src/domain/variants/variants-stacked-area-tile/index.ts
