+++ packages/app/src/domain/variants/variants-stacked-area-tile/variants-stacked-area-tile.tsx
